# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iRestora POS API",
        "description": "API Documentation untuk Sistem POS F&B iRestora",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and receive an access token",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "userpertama@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "passwordbaru"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful, returns accessToken"
                    },
                    "401": {
                        "description": "Invalid email or password"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get user details",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Returns the authenticated user"
                    },
                    "401": {
                        "description": "Authorization header missing or malformed."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current access token",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tables/{tableId}/cart/submit": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Submit cart customer sebagai Batch pesanan baru (order_batch)",
                "operationId": "123f9237efa9ea3548fa0a7233ea93db",
                "responses": {
                    "201": {
                        "description": "Batch dibuat dengan status pending_confirmation"
                    },
                    "404": {
                        "description": "Meja tidak ditemukan / QR token tidak valid"
                    },
                    "409": {
                        "description": "Outlet belum memiliki shift aktif, order tidak dapat dibuat"
                    }
                }
            }
        },
        "/api/orders/{orderId}/batches": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all batches for an order",
                "operationId": "f8a0667c8aca0b1afec7f8fb1c36d522",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of batches with items"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order-batches/{batchId}/confirm": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Kasir mengkonfirmasi atau menolak Batch pesanan dari customer",
                "operationId": "63c369e9489c50fa6ee36151221f951c",
                "responses": {
                    "200": {
                        "description": "Batch dikonfirmasi/ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/outlets": {
            "get": {
                "tags": [
                    "Outlets"
                ],
                "summary": "List outlets accessible by the authenticated user",
                "operationId": "937ef41b2ef902e1130076004219d7fc",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/outlets/{id}": {
            "get": {
                "tags": [
                    "Outlets"
                ],
                "summary": "Get outlet detail",
                "operationId": "26c768fb6a972e7fcfd8fed81b6faffd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Outlet detail"
                    },
                    "404": {
                        "description": "Outlet not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}